refactor[#567]  : hooks 폴더명 lib으로 변경
