+++ src/pages/map/page.tsx
@@ -1,5 +1,5 @@
 import { LocalMarkingListBottomSheet } from "@/widgets/map/ui";
-import { useMapQueryParams } from "@/features/map/hooks";
+import { useMapQueryParams } from "@/features/map/lib";
 import { MarkingPins } from "@/entities/map/ui";
 import { useGetBoundaryMarkerList } from "@/entities/marking/api";
