+++ src/widgets/map/ui/myMarkingListBottomSheet.tsx
@@ -1,7 +1,7 @@
 import { useLocation } from "react-router-dom";
 import { useMap } from "@vis.gl/react-google-maps";
 import { MarkingItem, MarkingItemSkeleton } from "@/widgets/marking/ui";
-import { useMapQueryParams } from "@/features/map/hooks";
+import { useMapQueryParams } from "@/features/map/lib";
 import { RangeFilter, SortTypeFilter } from "@/features/marking/ui";
 import type { Bounds, LatLng } from "@/entities/map/types/client";
 import {
