+++ src/widgets/map/ui/placeMarkingListBottomSheet.tsx
@@ -2,7 +2,7 @@ import { useLocation, useNavigate } from "react-router-dom";
 import { useQueryClient } from "@tanstack/react-query";
 import { useMap } from "@vis.gl/react-google-maps";
 import { MarkingItem, MarkingItemSkeleton } from "@/widgets/marking/ui";
-import { useMapQueryParams, usePlaceQueryParams } from "@/features/map/hooks";
+import { useMapQueryParams, usePlaceQueryParams } from "@/features/map/lib";
 import { useMapStore } from "@/features/map/store";
 import { SortTypeFilter } from "@/features/marking/ui";
 import type { Bounds, LatLng } from "@/entities/map/types/client";
